# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafecitos API WEB",
        "description": "API for Cafecitos",
        "contact": {
            "email": "devcitobo@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "d470cd500f8c07d41b81bc8be9c2fca1",
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order by ID",
                "operationId": "1f765bbdc4c98369ee76760efa0e8f78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status/{status}": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Change order status",
                "operationId": "27b83b44d7561ad3c2e04f578b5ce8dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/types": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all order types",
                "operationId": "c89d422c1ba00b73b1cf15aabc66c547",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/history": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "operationId": "6d58568debe17e577f967a67956891b9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "waiter_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lenPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/me/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get my orders",
                "operationId": "e8d4590704f737175ef96308aba43b6c",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lenPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/move-to-commissiong": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Move all paid orders to commissioning",
                "operationId": "c06414dae1d2a6e10e88ad2cdc9e9cb3",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/orders/move-to-processed": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Move all paid orders to processed",
                "operationId": "4fc064337a383614445b69ebb3721c61",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "csrf": {
                "type": "apiKey",
                "description": "Para obtener este token, primero asegúrate de haber iniciado sesión. Luego, ejecuta el endpoint GET /sanctum/csrf-cookie. Finalmente, abre las herramientas de desarrollador de tu navegador, busca la cookie 'XSRF-TOKEN', copia su valor y pégalo aquí. Este token es necesario para todas las peticiones que modifican datos (POST, PUT, PATCH, DELETE).",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}